Remove wasm code, fix issue where ctrl-c would not close SDL window
