+++ Makefile
@@ -14,7 +14,7 @@ else
 endif
 
 
-EMULATOR_OBJS = emulator.c disassembler.c wasm.c
+EMULATOR_OBJS = emulator.c disassembler.c linux.c
 DISASSEMBLER_OBJS = disassembler.c
 TEST_OBJS = tests/emulator.c tests/disassembler.c
 
@@ -39,10 +39,10 @@ $(OSX_EXECUTABLE):
 osx: $(OSX_EXECUTABLE)
 	$(OSX_RUN_CMD)
 
-# TODO: Need a proper build for emcc to be available
-wasm: $(EMULATOR_OBJS)
-	emcc disassembler.c emulator.c wasm.c -o index.html -s USE_PTHREADS=1 -s WASM=0 -s USE_SDL=2 --embed-file invaders.h --embed-file invaders.g --embed-file invaders.f --embed-file invaders.e
+# wasm: $(EMULATOR_OBJS)
+# 	# Must source ./emsdk_env.sh and then run this target until Bazel is implemented
+# 	emcc disassembler.c emulator.c wasm.c -o index.html -s USE_PTHREADS=1 -s WASM=0 -s USE_SDL=2 --embed-file invaders.h --embed-file invaders.g --embed-file invaders.f --embed-file invaders.e
 
 # TODO: Replace include and library paths when bazel build is implemented
 linux: $(EMULATOR_OBJS)
-	$(CC) $(CFLAGS) -I ../SDL-mirror/include -L ../SDL-mirror/build/build/.libs disassembler.c emulator.c wasm.c -l SDL2 -lpthread -o linux
+	$(CC) $(CFLAGS) -I ../SDL-mirror/include -L ../SDL-mirror/build/build/.libs disassembler.c emulator.c linux.c -l SDL2 -lpthread -o linux
