+++ linux.c
@@ -1,6 +1,12 @@
 #include "emulator.h"
 #include <sys/time.h>
 #include <unistd.h>
+#include <pthread.h>
+#include <assert.h>
+#include "SDL.h"
+
+#define RGB_ON  0xFFFFFFFF
+#define RGB_OFF 0xFF000000
 
 #define KEY_COIN 'c'
 #define KEY_P1_LEFT 'a'
@@ -9,9 +15,17 @@
 #define KEY_P1_START '1'
 #define KEY_PAUSE 'p'
 
-char* files[] = {"invaders.h", "invaders.g", "invaders.f", "invaders.e"};
 struct State8080* emulator;
 
+struct draw_screen_args {
+  unsigned char* game_memory;
+  SDL_Renderer* renderer;
+  SDL_Texture* texture;
+};
+
+char* files[] = {"invaders.h", "invaders.g", "invaders.f", "invaders.e"};
+
+int should_quit = 0;
 uint8_t in_port1;
 uint8_t shift0;
 uint8_t shift1;
@@ -70,17 +84,8 @@ void set_out_port(uint8_t port, uint8_t value)
     }
 }
 
-int main()
+void* emulator_thread(void* arg)
 {
-  emulator = calloc(sizeof(State8080), 1);
-  emulator->memory = malloc(16 * 0x1000);
-  read_rom_into_memory(emulator, 0, files, 4);
-  printf("Finished reading roms into memory\n");
-  uint32_t pixels[256 * 224 * 4];
-  memset(pixels, 0xFF0000, 256 * 224 * sizeof(uint32_t));
-  uint32_t white[256 * 224 *4];
-  memset(white, 0xFF0000, 256 * 224 * sizeof(uint32_t));
-
   double timeSinceInterrupt = 0.0;
   double lastTimer = 0.0;
   double nextInterrupt = 0.0;
@@ -146,7 +151,171 @@ int main()
     usleep(3000);
 
   }
-
-  return 0;
+  return NULL;
 }
 
+void events_thread(void *arg)
+{
+    struct draw_screen_args* args = (struct draw_screen_args *) arg;
+    unsigned char* fb = args->game_memory;
+    uint32_t buffer[256][224];
+    SDL_Texture* texture = args->texture;
+    SDL_Renderer* renderer = args->renderer;
+    SDL_Event event;
+    while (SDL_PollEvent(&event)) {
+      uint8_t ascii;
+      switch (event.type) {
+        case SDL_QUIT:
+            should_quit = 1;
+        case SDL_KEYDOWN:
+          ascii = (uint8_t) (event.key.keysym.sym & 0xff);
+          switch (ascii) {
+            case KEY_COIN:
+                in_port1 |= 0x1;
+                break;
+            case KEY_P1_LEFT:
+                in_port1 |= 0x20;
+                break;
+            case KEY_P1_RIGHT:
+                in_port1 |= 0x40;
+                break;
+            case KEY_P1_FIRE:
+                in_port1 |= 0x10;
+                break;
+            case KEY_P1_START:
+                in_port1 |= 0x04;
+                break;
+          }
+          break;
+        case SDL_KEYUP:
+          ascii = (unsigned char) (event.key.keysym.sym & 0xff);
+          switch (ascii) {
+            case KEY_COIN:
+                in_port1 &= ~0x1;
+                break;
+            case KEY_P1_LEFT:
+                in_port1 &= ~0x20;
+                break;
+            case KEY_P1_RIGHT:
+                in_port1 &= ~0x40;
+                break;
+            case KEY_P1_FIRE:
+                in_port1 &= ~0x10;
+                break;
+            case KEY_P1_START:
+                in_port1 &= ~0x04;
+                break;
+          }
+          break;
+      }
+    }
+    for (int row = 0; row < 256; row += 8) {
+        for (int col = 0; col < 224; col++) {
+
+            uint32_t rotated_row = col;
+            uint32_t rotated_col = 256 - row;
+            uint32_t offset = rotated_row == 0 ? rotated_col/8 : (rotated_row * 256 - (256 - rotated_col))/8;
+            assert(offset < 7168);
+            unsigned char pixel = fb[offset];
+
+            // Loop through the pixel's 8 bits.  Each bit corresponds to a 32 bit int
+            // in the SDL screen buffer.  Since the pixel is rotated counter clockwise
+            // by 90 degress the first bit in the pixel is actually the last row in the
+            // SDL screen buffer.
+            // 
+            // 1 byte pixel (before it's rotated)
+            //
+            // +---+---+---+---+---+---+---+---+
+            // |   |   |   |   |   |   |   |   |
+            // |   |   |   |   |   |   |   |   |
+            // +-+-+---+---+---+---+---+---+---+
+            //   ^
+            //   |
+            //   |
+            //   |
+            //   +
+            //
+            // The 8 32 bit ints for a given pixel
+            //          +----+
+            //  row     |    |
+            //          +----+
+            //  row + 1 |    |
+            //          +----+
+            //  row + 2 |    |
+            //          +----+
+            //  row + 3 |    |
+            //          +----+
+            //  row + 4 |    |
+            //          +----+
+            //  row + 5 |    |
+            //          +----+
+            //  row + 6 |    |
+            //          +----+
+            //   +----> |    |
+            //  row + 7 +----+
+            for (int i = 0; i < 8; i++) {
+                if ((pixel & (1 << (7 - i))) != 0)
+                    buffer[row + i][col] = RGB_ON;
+                else
+                    buffer[row + i][col] = RGB_OFF;
+            }
+        }
+    }
+    SDL_UpdateTexture(texture, NULL, buffer, 224 * sizeof(uint32_t));
+    SDL_RenderClear(renderer);
+    SDL_RenderCopy(renderer, texture, NULL, NULL);
+    SDL_RenderPresent(renderer);
+}
+
+int main()
+{
+    if (SDL_Init(SDL_INIT_VIDEO|SDL_INIT_AUDIO) != 0) {
+      SDL_Log("Unable to initialize SDL: %s", SDL_GetError());
+      return 1;
+    }
+    emulator = calloc(sizeof(State8080), 1);
+    emulator->memory = malloc(16 * 0x1000);
+    read_rom_into_memory(emulator, 0, files, 4);
+    printf("Finished reading roms into memory\n");
+
+    SDL_Window *window;
+    // Create an application window with the following settings:
+    window = SDL_CreateWindow(
+        "An SDL2 window",                  // window title
+        SDL_WINDOWPOS_UNDEFINED,           // initial x position
+        SDL_WINDOWPOS_UNDEFINED,           // initial y position
+        224,                               // width, in pixels
+        256,                               // height, in pixels
+        SDL_WINDOW_OPENGL                  // flags - see below
+    );
+
+    // Check that the window was successfully created
+    if (window == NULL) {
+        // In the case that the window could not be made...
+        printf("Could not create window: %s\n", SDL_GetError());
+        return 1;
+    }
+    SDL_SetWindowTitle(window, "Space Invaders");
+    pthread_t emu_thread;
+    int rc = pthread_create(&emu_thread, NULL, emulator_thread, NULL);
+    if (rc != 0) {
+      printf("Could not create pthread, exited with code %d", rc);
+    }
+    SDL_Texture* texture;
+    SDL_Renderer* renderer;
+    renderer = SDL_CreateRenderer(window, -1, SDL_RENDERER_PRESENTVSYNC);
+    texture = SDL_CreateTexture(renderer,
+      SDL_PIXELFORMAT_ARGB8888, SDL_TEXTUREACCESS_STATIC, 224, 256);
+
+    struct draw_screen_args args = {
+      .game_memory = &emulator->memory[0x2400],
+      .renderer    = renderer,
+      .texture     = texture,
+    };
+    while (! should_quit) {
+      events_thread(&args);
+      usleep(16000);
+    }
+
+    return 0;
+}
